Refactor loss functions and finally fix the cross-entropy loss implementation.

+++ SimpleAi.UI/Usings.cs
@@ -1,1 +1,1 @@
-global using NumberTypeT = float;
+global using NumberTypeT = double;
